# monolith/static/swagger.json
{
  "consumes": [
    "application/json",
    "multipart/form-data"
  ],
  "produces": [
    "application/json"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Documentation of auto trade API.\n\nSchemes:\nhttp\nhttps",
    "title": "infoblog.",
    "version": "1.0.0"
  },
  "basePath": "/",
  "paths": {
...
    "/api/1/exchange/balance": {
      "get": {
        "security": [
          {
            "Bearer": [
              "[]"
            ]
          }
        ],
        "tags": [
          "exchange"
        ],
        "summary": "Проверка баланса пользователя.",
        "operationId": "exchangeBalanceAccountRequest",
        "responses": {
          "200": {
            "$ref": "#/responses/exchangeBalanceAccountResponse"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "BalanceDTO": {
      "description": "go:generate easytags $GOFILE json,db,db_ops,db_type,db_default,mapper",
      "type": "object",
      "properties": {
        "amount": {
          "type": "string",
          "x-go-name": "Amount"
        },
        "currency": {
          "type": "string",
          "x-go-name": "Currency"
        },
        "locked": {
          "type": "string",
          "x-go-name": "Locked"
        }
      },
      "x-go-package": "studentgit.kata.academy/eazzyearn/students/mono/monolith/internal/models"
    },
    "BalanceInfo": {
      "type": "object",
      "properties": {
        "Balances": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/BalanceDTO"
          }
        },
        "Message": {
          "type": "string"
        }
      },
      "x-go-package": "studentgit.kata.academy/eazzyearn/students/mono/monolith/internal/modules/exchange_client/controller"
    },
...
    "GetAccountBalanceOut": {
      "type": "object",
      "properties": {
        "DataMargin": {
          "$ref": "#/definitions/BalanceInfo"
        },
        "DataSpot": {
          "$ref": "#/definitions/BalanceInfo"
        },
        "ErrorCode": {
          "type": "integer",
          "format": "int64"
        },
        "Success": {
          "type": "boolean"
        }
      },
      "x-go-package": "studentgit.kata.academy/eazzyearn/students/mono/monolith/internal/modules/exchange_client/controller"
    },
...
  },
  "responses": {
...
    "exchangeBalanceAccountResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/GetAccountBalanceOut"
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}